+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix [#85](https://github.com/zpix1/yt-anti-translate/issues/85) Video Titles inside of a Playlist are not untranslated on mobile
 - Fix [#86](https://github.com/zpix1/yt-anti-translate/issues/86) Channel Featured video title not untranslated on Mobile
+- Fix [#87](https://github.com/zpix1/yt-anti-translate/issues/87) Channel name above description not untranslated on Mobile
+- Fix issue where mobile description could be "untranslated" as the currently playing advert
 
 ## [1.19.5] - 2025-07-26